tools/EpicBanner/requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
cloudscraper==1.2.71
idna==3.11
pyparsing==3.3.2
python-steamgriddb==1.0.5
requests==2.32.5
requests-toolbelt==1.0.0
urllib3==2.6.3
python-dotenv==1.0.1
